1) KMail developers talk about unifying some features with KNode. Marc and one KNode developers say that they will work on KReaderWin to unify it, Zack says that he will write a common composer for both. 2) Zack doesn't have time to work on it. 3) Ralf tells Zack on Sunday that it's a big usability problem, Lars says it would be cool, people are getting excited. 4) 10 hours of coding later Zack introduces a first rough draft of the Komposer framework. A common composer for all email sending applications. I even managed to squeeze some documentation in it.
Just to make sure people understand: this is not editor!. It will not give anything html editing. Don is working
on html composing. This a framework to use all KDE text editors! So far only Krichtextedit is
used I'll port KTextEditor interface and hopefully the current KMail editor with Don to it.
This is a complete and extensible composer framework. (to be honest I'm not yet adding the
composer specific widgets to it as I'm working on some nice To, CC, BCC widgets). Feel free to test
and start hacking on it.

svn path=/trunk/kdepim/libkdepim/; revision=245243