Fix the bug in the KOrganizer preferences dialog that the default settings were immediately applied and not only when pressing ok or apply.
This was a tough one. KPrefsDialog/Module which is used by KOrganizer operated
directly on the public variables of the KConfigSkeleton and didn't use the API
so Waldos trick with the "used API" didn't work. Fixed it by changing
KPrefsDialog/Module to work on the KConfigSkeletonItems instead of the
variables. This made some other changes in the apps using KPrefsDialog necessary.

svn path=/trunk/kdepim/libkdepim/; revision=257037