This rather large commit is mostly made up of unintentional indentation changes.  Sorry.
The useful bits update KIMProxy to be able to track multiple IM clients (so if you use Kopete for ICQ, and Konversation for IRC, if Konversation implements KIMIface, you can get information pooled from both.

We now track IM app launches and update status as soon as new information sources are available.

Information is cached as a sideeffect of the above - some things are a bit faster.

As well as that it allows you to launch your preferred IM when it's not running by clicking the presence field in a KMail fancy header.

There will be some flicker due to brute force updating the message view on app start or exit, but this will be an infrequent event unless I've messed up and everything crashes every two minutes.

svn path=/trunk/kdepim/libkdepim/; revision=313909