Don't create bogus additional completion entries if the name is already of the form "Lastname, Firstname", as in the branch. Additionally clean up the code some.
One could maybe construct virtual entries of the form "Foo Bar" from the
original "Bar, Foo" to make completion on the first name possible in these
cases, but I have a feeling that this is not the right place to do that
kind of email parsing. Ideally the completion fields should be available
individually from kabc and ldap, probably.

svn path=/trunk/kdepim/libkdepim/; revision=396354