Finally, the new recurrence code for libkcal. It encapsulates the whole recurrence functionality (rrule, ex rules, recurrence date/times, and exceptions date/times) into one classe. The API stayed as close to the old one, but this new code is much more powerful than the old one (although you won't see a difference in the applications yet; But once the library code works, we can make use of this extended functionality).
Not everything is fully finished (in other words: it's not perfect yet), so expect a few rough edges, which will be smoothed out before the release... There shouldn't be any data loss, but korganizer or kalarm might behave strange. If you experience anything like this, please tell me!

A long discussion (including a description of the changes) with David Jarvie went on on kde-pim:
http://lists.kde.org/?l=kde-pim&m=112017476307741&w=2
http://lists.kde.org/?t=112017493900001&r=1&w=2


CCBUG:98793
CCBUG:87331
CCMAIL: kde-pim@kde.org


svn path=/trunk/KDE/kdepim/libkdepim/; revision=437527