Update to current working version.
At this stage, we can connect to the exchange server,
get the folder list (including any subfolders, etc),
and build a mail message that looks roughly right.

There is still a long way to go though, to complete
the mail buildup (to make it look like a proper 
multipart MIME message), and to add support for the 
other things that exchange can provide and akonadi
supports (like appointments, tasks, etc).

svn path=/trunk/KDE/kdepim/akonadi/; revision=667842