Moved the base for KPIM::FolderTreeWidget to its own files since it provides very common functionality. Named it KPIM::TreeWidget and used also as base of KMMimePartTree in kmail to maximize code reusal. Others might take advantage of it.
Avoid using numeric constants for column indexes in KMMimePartTree.
Use descriptive preprocessor constants instead.

Implemented sorting in KMMimePartTree: it's useful when many attachments
are present and one wants only the ones of a certain type.
By default the sort indicator is not shown and we keep the sort insertion
order. We also revert to this order each time the view is filled.

Some minor additions to documentation

Source cleanup: inverted the order of a couple of functions.


svn path=/trunk/KDE/kdepim/libkdepim/; revision=803140