big patch which makes the reduced link interface work for kdepimlibs and kdepim:
* Use target_link_libraries(... LINK_INTERFACE_LIBRARIES ...) to define the "link interface" of the libs in kdepimlibs.
To enable this, set the option KDE4_ENABLE_EXPERIMENTAL_LIB_EXPORT to on using "make edit_cache". Then
apps linking to libraries will only link to the library itself and to the libraries specified using
LINK_INTERFACE_LIBRARIES. This gives less dependencies for packagers, after startup and I think there were more advantages.

* To make this work, export_library_dependencies() may not be used anymore, but instead the library targets have 
to be "exported", so they can be imported again later by other projects and this way bring their link interface with them.
So now all library targets from kdepimlibs are exported in the set kdepimlibsLibraryTargets. 

* To be able to give the developer some useful error message, KdepimLibsDependencies.cmake now errors out
and tells the developer to update kdelibs, and the install location information is now in KdepimLibsInformation.cmake,
which tells the developer to update kdepimlibs if it's not recent enough.

So: you need now an up-to-date version of kdepimlibs and kdesupport !
If that's not acceptable let me know and I'll add some workaround.

* Search the xsl file in the install location of the already installed kdelibs, not in the current install location, which may be in a completely different place.

* Use the correct install location (KDEPIMLIBS_INSTALL_DIR) in KdepimLibsInformation.cmake instead of KDE4_INSTALL_DIR, which is the one for kdelibs, but not for kdepimlibs.

* more minor things I don't remember right now

* Do you actually have a version number for kdepimlibs ? It would be easy to add via KdepimLibsInformation.cmake

Alex


CCMAIL: kdepim@kde.org
CCMAIL: winter@kde.org


svn path=/trunk/KDE/kdepim/akonadi/; revision=891439