Make the MDA work in the following situations: * dispatch messages when they appear in the outbox * dispatch messages on startup * don't dispatch while offline; begin dispatching when online
The MDA is now using a ProxyModel to filter the items in the outbox.  This model allows only valid items that should be dispatched at the current time.

There are a number of HACKs related to how the ItemFilterProxyModel works, and how the SendJob works.  I am also a little worried about speed.  A possible improvement is making SendJob an Akonadi::Job as opposed to a KCompositeJob, but that requires that Akonadi::Job be extended to support non-Akonadi subjobs.


svn path=/trunk/playground/pim/maildispatcheragent/; revision=968583