What about a filter editor based on QToolBox ?
It could fit the "recusive" nature of the new filter
model quite nicely as each rule box could contain a whole
rule list editor (with just a small indent).

The solution with the list box on the left side, instead,
eats a lot of space on the left and will probably make
the UI unusable with more than 2 levels.

The QToolBox, with some tweaks, would also be very kool :)


svn path=/trunk/playground/pim/akonadi/; revision=969830