Remove the automatic resource name setting.
I don't think it is very user friendly to have a filename as the name of the root
collection.

Also, this overwrites the settings from firstrun.cpp. Now one sees "Personal Calendar"
instead of "std.ics" for the default installation.

Volker, please review and tell me if I broke something with this.

svn path=/branches/work/akonadi-ports/kdepim/akonadi/; revision=1030651