Put in place a better logic for the management of IDLE within the resource. Mainly only triggers the collection resync if recent or total count of messages changed from the last update we got in this session. And since we have to start with a select we have numbers to start with.
Why did it get modified like that? Easy, previously we used recent as if
it was like unseen, while it's not. And the recent information doesn't
update for our IdleJob because it's listening on another session than
the main one which actually pump the data. Since the server might insulate
the clients from each others in such a case, the logic can't be based on
absolute numbers, but only on relative ones (as in "did it change from
last time?")

Looks like on of those "oh it'll be easy to fix" gone mad.


svn path=/trunk/KDE/kdepim/akonadi/; revision=1036055