We don't use the autosave timer anymore but schedule a write after changes are made to the resource. If new changes are made before the write is performed these will be scheduled before the write by the scheduler.
Also a minor fix in singlefileresource.h. When calling readLocalFile()
it cannot be garuanteed that the call succeeds (e.g. the file has
invalid content). So now readLocalFile returns a bool and the status
of the resource is set appropriate when the file could not be read.

TODO: Remove the autosave timer from the settings(dialog)

svn path=/branches/work/akonadi-ports/kdepim/akonadi/; revision=1045642