Treat remote ID change on a non-existing maildir as creating it.
During initial startup with an empty configuration, the Local Folders
resource will create 5 collections in rapid succession. It appears that
some of those might get their RID update coming into the resource before
they have been actually added. If that happens, treat the update for the
nonexisting collection as a creation of that maildir. This makes the
from scratch initial startup create the local folder reliably. I'm not
sure if the out of order communication between server and resource is
correct or not, but this seems like it should be safe.

svn path=/trunk/KDE/kdepim/runtime/; revision=1097296