Remove the "Store Password" checkbox.
Instead, assume that the user wants to store the password if he enters
it into the password edit.
This makes the UI more consistent with the IMAP settings.

The "store password" setting is still in the kcfg, so that the resource
doesn't needlessly open the wallet when the password is not stored.
In this case, the user is asked for the password with a popup.

svn path=/trunk/KDE/kdepim/runtime/; revision=1110605