pop3 resource: when connection fails display a visible notification with an option to view the settings dialog.
Before this commit, from the user's perspective it would silently
fail, though there is a debug message printed. This commit does not
break string freeze, because this code is directly copied from the
imap resource.

Technically, this dialog is misleading because the login can fail even
when the credentials are valid, such as when the auth/encryption
settings are incorrect.

svn path=/trunk/KDE/kdepim/runtime/; revision=1134715