Simplify sync logic.
Remove the distinction between incremental and non-incremental listing
since currently we will always list the full folder anyhow, to ensure
we get flag changes and deletions on the server. Per default, only
list flags and save listing of the full payload for the case of new
mails from the server (or the initial sync). Speeds up mail checks
hopefully quite a bit.

svn path=/trunk/KDE/kdepim/runtime/; revision=1139250