Remove extra i18n methods from kde plugin for qml
Making the slot's arguments be a QScriptValue var we
are able to access the QML's engine and then have access
to the context where the code is running.

With the context in hands, we are able to know the number
of arguments the js method was called and are able to parse
each one of them, calling the proper c++ overloaded method.

This is a workaround for not being able to directly get the
context (or engine) from QML. Another option would be to make
the plugin inherit QScriptable and thus we would have access
to the environment where this object was exported to, but this
doesn't work as QML doesn't use QtScript's mechanisms to create
and call methods.

svn path=/trunk/KDE/kdepim/runtime/; revision=1139674