We should keep the MouseArea as the logic in Qt seems to be right
A QDeclarativeItem defaults to not receive mouse events so not all
the items receive them (just the ones that are supposed to handle
them). There is no way to keep this behavior and just change something
on QDeclarativeMouseArea as the event itself never arrives to it's
child due to QGV's filtering (the item does not accept mouse events).

We could replace this Item{} for any other dummy class that would
properly handle the events but it's just easier to use a MouseArea
that doesn't handle the events itself, but properly propagates them.

svn path=/trunk/KDE/kdepim/runtime/; revision=1161139