Avoid unnecessary collection tree syncs.
When a folder is modified that contains email, but is a parent of kolab
groupware folders (such as the INBOX), treat it as the structural node
in the kolab collection tree that it is, instead of a suddenly appeared
new kolab folder. To that end, don't just check if the folder is the
folder-type annotation, also make sure it's not "mail", to tell kolab
folders from structural ones.

svn path=/branches/KDE/4.5/kdepim/runtime/; revision=1169244