Let the MDA re-try a few times to create or get the outbox.
In the case of a fully clean startup, with no configuration, the
MDA and mail app will race for creation of the outbox, etc. Since
there is no notification between the processes about changes to the
special collections, we now re-try a few times, with a back off,
and if that still doesn't yield a working outbox, give up.

svn path=/trunk/KDE/kdepim/runtime/; revision=1207551