Set the index to dirty when saving the settings for the agent.
Previously we only re-indexed when the compat level changed, but
if one throws away the index, and the feeder, and later re-enables
it, the compat level of the index hasn't been reset. An option would
be to reduce the compat level by one, upon saving changes, but let's
be clean and simply keep an explicit boolean to force re-indexing
on config changes. As a side effect, this makes sure re-indexing
happens when the feeder is removed and re-added.