Only call task management functions when in task context.
The readFile and writeFile methods can be called from within a
sceduler task, in which case they need to cancelTask on error and emit
taskDone in the end, or they can be called directly. Since we don't want
to randomly cancel ongoing other taks, in that situation, make it
explicit which kind of processing we're in.