Fix compilation with srcdir == builddir
Both "." and "notes" would generate a settings.h, and that would break
the #include "settings.h" in icalresourcebase.h which is used from both dirs,
since it would then always find the one in "." due to the double-quotes-include.
Moving the shared code to a shared subdir fixes that, it moves it away from
both generated settings.h files.