Adjusting item fetch test for new maildir lib capabiltities.
The maildir library got the capability to retrieve certain flags
from the mail file names, i.e. the status flags maildir stores as
part of each entry's file name.

The fetch test has been adapted to reflect that by checking for the
flags that are now present, especially in case were the index is not
available or outdated