Fix modification checks in mixed maildir store and associated tests.
With the support for changing flags on maildir mails, the check for
successfully doing nothing has to be moved after the check of
remote ID validity.

Adapting the unit test to specify which part(s) of the item has been
changed to pass the store's check for necessity when required.