Ignore items with empty remote ids here.
Items with empty remote id are those not yet stored by the resource,
ie. we still have pending change replays for those. So, if we do a sync
during that period (which could happen since the introduction of the
file system watcher), we would remove a random one of those (only one
since the map would only contain one without a remote id). This mostly
affects the POP3 usage pattern.