Auto commit on transaction is off now, so always commit if we have one
Automatic commit of transaction has been turned off in
2667289521c2ba4526434430e1db2d4b1a3bfa2b but explicit commit
only happens when timestamp information of the collection needs to be
modified.
This is usually the case if there is any change, but just to be sure
always commit if we reach the end of the job's item processing and
there is a transaction (indicating that there has been some change).