Update migration version information on success
KMail and the migrator itself check the "Version" entry in group
"Migration" and compare its value with the one from "TargetVersion".

If version is less then target version, the migrator is run again.

This makes sure the version value is updated to the current target value
whenever the migrator runs without error.