Fix livelock in feeder.
The feeder somehow managed to lock itself (probably caused by the block var).
This patch counts the running jobs instead of the boolean block variable which seems less fragile, because it's not necessary to know all possible entry/exit points of the queue.
The timeout after error was removed because the dms still doesn't merge all items properly, but the timeout is no longer needed because dbus doesn't time out anymore (dbus timeout increased).
This greatly improves the performance when there are a lot of indexing errors.