Save the indexCompatLevel with each item.
This allows us to check if the items need updating in case the indexer
improved.
We now have a 2-step initial update: The initial update is done once
all collections are indexed. This is written to the config as before.
However, if the index compat level changes the initial indexing is
restarted. But in order to not re-index every single item between
restarts (in case the initial update did not finish) we check for the
compat level used on the item and only update those which have an old
one.