Don't remove resources before updating them.
We can simply overwrite the properties. Before also PIMO things were removed because they are marked as subresource of the item,
(which is correct), but therefore linked resources were lost on every update.
This patch resolves that, with the drawback that plugins need to take care of outdated subresourcecs themselves.