Handle location specified by path or URL. Handle missing config file.
1) The collection remote ID contains the calendar file location, but
if the config file contains the URL instead of path, the config wasn't
recognised and the resource failed to initialise correctly, leaving a
broken resource. This commit makes the resource handle either path or
URL in the config.

2) If the resource config file akonadi_kalarm_resource_*rc was
missing, the resource was recreated but made broken due to the
missing config. This commit makes the resource recreate the config
from the collection properties if at initialisation the path is
missing from the config.