Handle location specified by path or URL. Handle missing config file.
If the resource config file akonadi_kalarm_dir_resource_*rc was
missing, the resource was recreated but made broken due to the
missing config. This commit makes the resource recreate the config
from the collection properties if at initialisation the path is
missing from the config.

Make the resource hangle either path or URL in the config.

Backport of 0ff07b36564a7103d5ede92a97909a8fb24ebb1a (master).