Fetch Google contacts photos in batches to avoid reaching quota
The number of Google API requests per user per second is limited.
Fetching high amount of contact photos (usually at initial sync)
almost always hits the quota and so some contact photos are not
synced.

This commit queues the requests and sends them in batches. When we hit
the quota, the interval between batches is increased and the failed
requests are queued again.

BUG: 303749
FIXED-IN: 4.9.3