Slightly optimize memory usage and make sure the used memory is released ASAP.
Most memory is being used by RecursiveItemFetchJob, so we make sure that
it gets the chance to delete itself immediately after it's not required
anymore. Further some memory can be saved by storing Id's only. The
lists we're dealing here with can grow as large as all items in your
akonadi db.

Memory usage could be further optimized by re-implementing
RecursiveItemFetchJob and continuously freeing the jobs after we're done
with a collection.