Keep custom name/icon of resource collection on collection sync
To ensure that the user modifies the display name of the resource
collection and not the collection name itself, we have to preset
the display name in the EntityDisplayAttribute already.

Additionally the clearCache() call must be removed, it's not needed
if we use the directory path as remoteId, because then the collection
sync algorithm will cleanup the collection with the old remoteId.

BUG: 314446
FIXED-IN: 4.10.2