Nepomuk Feeder Cleaner: Fetch the Nepomuk URI when fetching all the indexed items
This is done because it is easier for Nepomuk to remove all the items if
we provide it with the nepomuk uris, instead of it having to query for
each one separately.

Also, certain stale data only has the akonadiItemId and not the nie:url
set, so that data cannot be removed cause the nepomuk dms internally
only maps the nie:url to the uri.