Nepomuk Feeder: Do not emit events/change the status on idle
1. The message indicates that the system is busy and indexing has been
suspended when that is clearly not the case. The Indexing has just been
throttled. This is the main reason people wonder why virtuoso is
consuming thier cpu even though nothing is indexing.

2. This intefers with the other status messages. Say the indexing has
finished, and the "Indexing Completed" message is shown. When the system
becomes idle that message will change, which is fine, but then it
changes back to "System busy, indexing suspended", which is so wrong.

I rather not that these status messages emitted at all. We already have
decent status messages emitted by the FeederQueue.