Initial support of dav resource for Accounts
Since we do not want to port the whole resource to Accounts I have
taken the following approach:

When the configuration is reloaded:
	-Remove any remoteUrl that belong to a disabled service
	-Create new remoteUrl for newly enabled services

Those urls that belong to Accounts have $accounts$ as the username so
we will be able to continue having the same functionality we had until
now, plus Accounts.

In the future I'd like to port the resource completely but future me
will take care of that.