Display KNotification and KStatusNotifier for Facebook notifications
Whenever the user receives new Facebook notification, it's displayed in
KStatusNotifierItem, which also includes menu with direct links to the
notifications as well as an option to mark all notifications as read (on
the server).

For every new Facebook notification there is KNotification displayed as
well. If there are more than one Facebook notifications, the newest one
is displayed with "...and N more" appended.

The configuration part of the resource contains a setting to turn this
off.

REVIEW: 109825
DIGEST