NepomukFeeder: Remove the concept of Batch Indexing and checking
The feeder has this concept of checking every hour if it had not indexed
any emails. This resulted in huge CPU loads every hour because the
process to calculate which emails need to be indexed is very cpu
intensive on both the akonadi and virtuoso side.

Additionally, it had this concept of batches. If more than 10 emails
were receieved in a span of 10 seconds it would not queue them up and
run this heavy email check in 30 minutes.

Using random intervals such as 30 and 60 minutes is bad since the user
has no idea what is going on. Specially if they try to search for some
email during that time. Additionally, they can understand if Nepomuk is
taking some time indexing stuff when you get new emails, but not why it
is suddenly taking a LOT of their CPU after 30 minutes. They loose
context.