NepomukFeeder: Schedule the cleaning jobs
Do not just run them when required. That leads to a lot of jobs running
in parallel which is something we do not want. Instead, schedule them
via the scheduler.

Currently, they are scheduled when we have nothing left to index, since
who cares about unindexed stuff! It won't even be shown to the user
since the akonadi url stat will fail.