Nepomuk Feeder: Remove special Collection handling code from Scheduler
Removed the code to request all the items in a collection and index
them. From now on when a collection is added to the scheduler, only the
collection is indexed. Not the items in the collection.

This makes sense because we should be receiving a signal for each item
in the collection when the collection is added and populated. It makes
little sense to index the items twice.

This code used to exist for manual indexing of a collection. This is not
the correct place for that. There is a helper utility can be used for
develoeprs. Normal users should never need to force reindexing of a
collection.