Nepomuk Feeder: tell the storage service to merge duplicates in the graph
This is a slightly costly operation, but it is required. Lots of emails
seem to often have the same contact in both the from and to. If the
contact doesn't already exist in the database, then 2 copies of it are
made. This slows down future email indexing related to that contact
cause the duplicate detection code always needs to choose one.