Make kolab resource use the same tasks icon as korganizer's views.
The icons are almost identical, this is just a name change.

Month/Agenda view items show the collection's icon, and also the tasks
icon in case of to-dos. When korg sees that both icons are the same
it will only display one.

Because kolab was using an icon with a different name, korg was showing
two identical tasks icons on each item.