Add file monitoring to iCalDir and vCardDir resources and move common code to shrade base class
Similarily to iCal and vCard resources, the iCalDir and vCardDir resources
are essentially identical. This patch moves all the code to a template-based
DirResource class, making VCardDirResource and ICalDirResource specializations
of the DirResource class, implementing only loading and saving of payload
from/to file. QObject-related stuff is implemented in DirResourceBase class.

The shared code now also supports watching the directory (via KDirWatch)
for changes and will automatically add, update or remove entries from
Akonadi when a file is added, changed or removed in the resource's data
directory.