Added api function: - setSelectedDay(int year, int month, int day); - in QML the model: selectedDayModel is added.
This was needed for a second model that shows the entries for the selected day. This can potentially also be used as popup when you hover a day.

This commit kinda makes the CalendarData component useless as it is in it's current state since the "Calendar" component can do everything that "CalendarData" can.
In the C++ side it is being used to fill the above mentioned second model.

Renamed:
- in QML: "daysModel" to "model" because the name was confusing.