GID migration.
A new MigratorBase as interface for migrators.

I didn't adapt the existing KMigratorBase as it taylored towards one
migrator per resource and I found a couple of other incompatibilities so
I figured I rather make something clean from scratch.

It should be easy enough to adapt KMigratorBase to be based on top
MigratorBase if we still need that code.

REVIEW: 112635