Migration agent that schedules various migrators and exposes an interface for status and control.
This does not yet include a plugin interface for making plugging in new
migrators plugin based. That would be easy enough to add but only makes
sense if we push the MigratorBase interface to kdepimlibs along with a
dbus control interface, so i.e. applications such as kmail can install
new migration plugins and control them via the dbus interface.

The agent automatically starts migration jobs that have
shouldAutostart() enabled and exposes an interface in it's configuration
dialog to start/stop migrators. The interface is yet very basic and is
also no necessarily in the most convenient place. We should likely have
some central place for controlling background processes or simply ensure
the automatic processing is handled well enough.

REVIEW: 112636