Don't bother IMAP servers with not supported flags
Created a new method ResourceTask::fromAkonadiToSupportedImapFlags() to convert
Akonadi to IMAP flags supported by the server and use it everywhere instead of 
fromAkonadiFlags(). 

fromAkonadiFlags() was made private to make it impossible to call it by accident.

BUG:278082
FIXED-IN: 4.11.3
REVIEW:112724

CCBUG:287356

This also reduces the risk to trigger 287356 as the insert on the IMAP server
doesn't fail due to unsupported IMAP flags.