Instead of being spontaneous we just keep the metadata up to date.
The whole "spontaneous" idea was a workaround around the problem created by not updating
the metadata. As this workaround has been removed and we're commited to have up-to date metadata,
there is now no point in this workaround anymore.

This probably also removes the need for the whole retrieveCollectionAttirbutes API.