Remove the Microblog resource
Twitter and Identica APIs are now different, plus the Twitter API
has changed a bit in past, so the resource does not work properly
anymore. Also we don't really have any application to show the stuff.
There's a new microblog resource living in a scratch repo, in case
we want that, we can just nag mck182 to finish and merge it.