Try harder to avoid callbacks in half destructed state.
After the destructor callbacks will result in undefined behaviour, we destory everything
that might call us back. Aditionally, all the logout stuff is moved to aboutToQuit, since in the destructor it's really too late for stuff like that.

This will hopefully fix some of the crashes that show up when shutting down the imap resource.