IMAP-Resource: Async dialogs.
Async password request dialog that is aborted after disconnecting.

This properly aborts the dialog after the connection attempt is no longer relevant,
and avoids the stacking of dialogs.

The configuration dialog is also made async to avoid similar problems.

With this change the dialog appears at most once, and vanishes if the network is disconnected again.

BUG: 306955
BUG: 319171