Disable CONDSTORE when using Gmail
Gmail's implementation of CONDSTORE is weird and breaks changes
synchronization of folders other than INBOX. Until we can figure
it out (or untile a Gmail Resource is written), we better disable
CONDSTORE when using Gmail, so that people can get to their emails.

This means that mail check with Gmail is going to be slow again, but
no other functionality is lost.

Approved-By: Christian Mollekopf