Kolab-Resource: New Kolab Resource that subclasses the imap resource (not a proxy anymore)
The KolabRetrieveCollectionsTask is a refactored copy, that should eventually
replace the imap resources version. It's kept separate for now until
it's stable.

Instead of separating the groupware/non-groupware trees we mix them.
This way retrieving the collections remains a single IMAP command, and not one
per folder, which would be prohibitively expensive with thousands of folders. Instead we
always retrieve the metadata before syncing the folder.
This also has the advantage that the kolab resource code stays closer to the imap code.