IMAP-Resource: Refactoring of retrieveitemstask
This introduces the following changes:
* Batch wise fetch to not overload akonadi
* Proper progress reporting through setTotalItems
* Make it clear that incremental/non-incremental is a mode of the task that
cannot be changed.
* Removed the processed mailbox hack
* Introduced a check for inconsistent local cache state (since we rely on this to be correct)
* Made sure the local collection is actually cleared when empty on the server
* Made sure changesince is only used in inremental mode
* Some minor refactorings for code clarity
* Fixed "retrieve items without body" mode, which just worked by chance
and would delete all your local mail if it once didn't end up in
incremental model.
* Disabled "retrieve items without body" mode. This should be called
explicitly by the resource when required. It's very expensive and certainly
shouldn't be executed on every sync. The only reason it's still in the
same job is for codesharing (it's really entirely unrelated otherwise).