Kill Akonadi Tray applet in Frameworks
The applet does nothing useful except for annoying people with notifications
about error messages. It's main functionality, backup and restore of Akonadi
does not work properly with external MySQL and it does not support SQLITE db
at all.

If we want a backup/restore, we should have this implemented in Akonadi server
and we should have a proper application (KCM?) for that.