IMAP-Resource: Make sure we really detatch the collection.
Akonadi::Collection silently shared attributes with other copies because it
doesn't properly detatch, because Entity::attribute is const.
This resultet in uidnext getting modified even if the task was canceled,
because ResourceState::cancelTask modified the original copy of the collection
(which erroneously received the updated attributes due to the detatch bug).

We now force a detatch to avoid nasty surprises.