IMAP-Resource: Don't populate the resource with flags only.
Catch another case that can easily be triggered by clearing the cache,
while leaving the attribute intact. We tried to do an incremental sync in that
case, and the resource would populate the whole collection with flags only.
We now detect that something is off with the messagecount, and do a full sync instead.