Sort by filename, which is RemoteId, instead of item id to make test pass.
Review request 118441 introduced sorting of items for ItemFetchJob
testcase as due to an performance optimization mixedmaildir resource
does not sort anymore. This fixed the testcase for me on BTRFS.

However on the Jenkins server the test case was still not fixed and
Kevin pointed out that qSort may sort by item id instead of RemoteId.
The test relies on sorting by filename, which is basically the RemoteId.

CCBUG: 334218

REVIEW: 118619