Gmail Resource: Implement custom Cyrus-SASL plugin for XOAUTH2 mech
Cyrus-SASL that we use in KIMAP does not support XOAUTH2 and will
refuse to authenticate us via this mech. This is a very simple
implementation of client-side plugin for Cyrus-SASL that adds
support for XOAUTH2 authentication mechanism.

In future, this should probably go to KIMAP, because KIMAP's XOAUTH2
support is useless withouth this.