Gmail resource: ignore the [Gmail]/Chats folder for now
Chats is a special folder, it's content is not owned by 'All Mail' folder
and does not work as a label either. This breaks sync for some reason, so
we disable the folder for now.

The folder is not exposed via IMAP by default, it has to be explicitly
enabled in Gmail web interface, and it contains preprocessed hangout
logs in the form of emails, which I don't see much use for in KMail
anyway.

BUG: 337497
FIXED-IN: 4.13.90