Manually follow redirect when creating items
Needed for at least Zimbra that doesn't like
having items files named by anything else but
the incidence UID and that redirects. KIO
does not supply the username / password (as
it should), but we have no way to ask the user
for confirmation.

Many thanks to Benjamin Girault for the initial
debug, patch and testing.

BUG: 339634