Unbreak syncing on Kolab resources
triggerCollectionExtraInfoJobs was removed in 945a5cca9 but the
corresponding change never made it to the Kolab resource, thus it would
stop when trying to run a non existent task.

Notice that this may be a stop-gap solution (I don't know the code too
well) but seems to work.

CCMAIL: dvratil@redhat.com
CCMAIL: chrigi_1@fastmail.fm