Make BatchFetcher workaround MS Exchange "bugs"
If the result of UID SEARCH is too big Exchange just disconnects. To work around
this the UID SEARCH is splitted into smaller chunks (I choose 2000 arbitrarily).

First I wanted to do this for all servers but I've a server where UIDs go up to
1 billion which results in half a million SearchJobs. So I do the splitting only
for Exchange servers. I don't know if there is a better way to recognize Exchange
servers than to look at Session::serverGreeting()?

There are still issues with Exchange but at least for me it's usable again.


CCBUG:  339393
REVIEW: 120644