IMAP-Resource: Remove all uses of the timestamp attribute.
Updating this during every metadataretrievel resulted in a lot of updates,
in i.e. kmail, that currently are very expensive (due to the favoritesproxymodel
issue). Since it was only used in one place to determine wether the collection
is new, we can drop it entierly. I removed the knownMailboxes hack with it,
since we no longer use it anyways.