Fix GETANNOTATION for Other Users
If GETANNOTATION/METADATA is requestes for many folders, the server can
get stuck and the connection will timeout. To fix this, we ask only for
subfolders, when requesting for folders in "Other Users".

For searching we need to request all metadata for "Shared Folders".
"Shared Folders" is not a valid mailbox, so we can't request ACLs for
it. If no "Shared Folders" is available, job do not fail.

KOLAB: 3875