akonadi_rsrc_pop3: Skip fetching already-present msgs.
The POP3 resource does a pre-check when downloading messages from the
POP3 server to filter out messages we already have. This pre-check can
result in having no messages to download at all. Currently we still
request the POP3 kioslave to retrieve a message (without an id) which
leads to warnings.

This change advances directly to the next state instead, when there are
no messages that need to be downloaded.

REVIEW:124988