IMAP: use UIDNEXT instead of "*" as interval end when doing full mailbox resync
The opened interval is represented as -1, so when syncing with Microsoft
Exchange the qMin() comparision in the search-in-chunks workaround for
Exchange evaluates as qMin(X, 2^32-1), causing the BatcherFetcher to get
to a loop trying to SEARCH all chunks between 1 and 2^32-1. In practice
Exchange will drop connection at some point, so the folder sync is aborted
and no emails are ever synced.

With this patch we limit the interval end to reported UIDNEXT, so that the
SEARCH loop is finished when we reach the interval with UIDNEXT.

BUG: 351814
FIXED-IN: 15.08.2