Revert "Kolab-Resource: Ignore flag changes and don't accidentaly create mails."
This reverts commit ee89b1e57b001ce3533647198ba00db6cd7f9812.

Not a good idea at all. This breaks the regular sync where we retrieve
new items, and flags for the rest only, and then remove all items for
which we didn't retrieve flags.

A better fix has to not create email items even if only flags are
provided and the item doesn't locally exist yet.