KolabProxy: always synchronize underlying IMAP collection
When KolabProxy::retriveItems() is called the proxy only synchronizes
its state with the current state of the underlying IMAP collection but
does not sync the IMAP collection itself. This change triggers the sync
of the IMAP collection after the Kolab collection is synchronized.

This fixes "Update this folder" action in KMail and others which had
no effect otherwise on collections managed by the Kolab Proxy resource.
The only way to trigger the actual sync was via 'Check mail' button in
KMail, which directly called a DBus method on all email resources.