KolabProxy: workaround race condition between the Proxy and IMAP
This affects specific situation when the IMAP resource syncs a
changed item (for example an updated event). The ItemSync in IMAP
resource will create a new Item (with the changed content) and then
delete the old one.

The creation of the new Item is caught by the Kolab Proxy and it
goes on to check if an Item with same GID already exists. That happens
to be true in most cases because the Proxy gets to check for this
before the ItemSync in the Kolab Resource gets to delete the original
Item. However when the Proxy tries to update the existing Item it
fails to do so, because the Item has been deleted in the meantime by
the ItemSync in IMAP.

To workaround this issue the Proxy now does Create+Delete instead
of Modify to ensure we don't race with the IMAP resource about the
item.

This fixes changed events disappearing from KOrganizer after sync
and only re-appearing after next sync.